+++ flexiapi/app/Http/Controllers/Api/PhoneCountryController.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace App\Http\Controllers\Api;
+
+use App\Http\Controllers\Controller;
+use App\PhoneCountry;
+
+class PhoneCountryController extends Controller
+{
+    public function index()
+    {
+        return PhoneCountry::all();
+    }
+}