+++ flexiapi/routes/web.php
@@ -38,6 +38,7 @@ use App\Http\Controllers\Admin\AccountController as AdminAccountController;
 use App\Http\Controllers\Admin\AccountStatisticsController;
 use App\Http\Controllers\Admin\ContactsListController;
 use App\Http\Controllers\Admin\ContactsListContactController;
+use App\Http\Controllers\Admin\PhoneCountryController;
 use App\Http\Controllers\Admin\SipDomainController;
 use App\Http\Controllers\Admin\StatisticsController;
 use Illuminate\Support\Facades\Route;
@@ -157,6 +158,14 @@ Route::group(['middleware' => 'web_panel_enabled'], function () {
         Route::middleware(['auth.super_admin'])->group(function () {
             Route::resource('sip_domains', SipDomainController::class);
             Route::get('sip_domains/delete/{id}', 'Admin\SipDomainController@delete')->name('sip_domains.delete');
+
+            Route::name('phone_countries.')->controller(PhoneCountryController::class)->prefix('phone_countries')->group(function () {
+                Route::get('/', 'index')->name('index');
+                Route::get('/activate_all', 'activateAll')->name('activate_all');
+                Route::get('/deactivate_all', 'deactivateAll')->name('deactivate_all');
+                Route::get('/{code}/activate', 'activate')->name('activate');
+                Route::get('/{code}/deactivate', 'deactivate')->name('deactivate');
+            });
         });
 
         Route::name('statistics.')->controller(StatisticsController::class)->prefix('statistics')->group(function () {