+++ flexiapi/tests/Feature/ApiAccountPhoneChangeTest.php
@@ -20,8 +20,8 @@
 namespace Tests\Feature;
 
 use App\Account;
-use App\AccountCreationToken;
 use App\PhoneChangeCode;
+use App\PhoneCountry;
 use Tests\TestCase;
 
 class ApiAccountPhoneChangeTest extends TestCase
@@ -55,7 +55,7 @@ class ApiAccountPhoneChangeTest extends TestCase
 
         $this->keyAuthenticated($account)
             ->json($this->method, $this->route.'/request', [
-                'phone' => '+123123'
+                'phone' => '+33612312312'
             ])
             ->assertStatus(200);
 
@@ -69,6 +69,35 @@ class ApiAccountPhoneChangeTest extends TestCase
                 'code' => $account->phoneChangeCode->code
             ])
             ->assertStatus(410);
+    }
+
+    public function testCreatePhoneByCountry()
+    {
+        $account = Account::factory()->withConsumedAccountCreationToken()->create();
+        $account->generateApiKey();
+
+        $frenchPhoneNumber = '+33612121212';
+        $dutchPhoneNumber = '+31612121212';
+
+        $this->keyAuthenticated($account)
+            ->json($this->method, $this->route.'/request', [
+                'phone' => $frenchPhoneNumber
+            ])
+            ->assertStatus(200);
+
+        $this->keyAuthenticated($account)
+            ->json($this->method, $this->route.'/request', [
+                'phone' => $dutchPhoneNumber
+            ])
+            ->assertJsonValidationErrors(['phone']);
+
+        PhoneCountry::where('code', 'NL')->update(['activated' => true]);
+
+        $this->keyAuthenticated($account)
+            ->json($this->method, $this->route.'/request', [
+                'phone' => $dutchPhoneNumber
+            ])
+            ->assertStatus(200);
     }
 
     public function testUnvalidatedAccount()